Renamed to PinchUtility
