+++ Assets/LeapMotion/Scripts/LeapHandController.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
 guid: 215a4d49fc705b74a9d3c5cbfa2c9601
-timeCreated: 1455681390
+timeCreated: 1456448226
 licenseType: Pro
 MonoImporter:
   serializedVersion: 2
   defaultReferences: []
-  executionOrder: -1000
+  executionOrder: -1002
   icon: {instanceID: 0}
   userData: 
   assetBundleName:
