+++ Assets/LeapMotion/Scripts/LeapProvider.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
 guid: 025cc0fa7b46aa541aba29d28d35ac09
-timeCreated: 1455681390
+timeCreated: 1456448226
 licenseType: Pro
 MonoImporter:
   serializedVersion: 2
   defaultReferences: []
-  executionOrder: -1001
+  executionOrder: -1003
   icon: {instanceID: 0}
   userData: 
   assetBundleName:
