+++ Assets/LeapMotionModules/PinchUtility/Examples/Materials/PinchLineMaterial.mat
