+++ Assets/LeapMotionModules/PinchUtility/Examples/Scripts/PinchDraw.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
 guid: f6f4cecb16a3e0d4db78a7f07e12721c
-timeCreated: 1455830864
+timeCreated: 1456448227
 licenseType: Pro
 MonoImporter:
   serializedVersion: 2
   defaultReferences: []
-  executionOrder: 0
+  executionOrder: -1000
   icon: {instanceID: 0}
   userData: 
   assetBundleName:
