+++ Assets/LeapMotionModules/PinchUtility/Examples/Shaders/LeapPinchLine.shader.meta
