+++ Assets/LeapMotionModules/PinchUtility/Scripts/LeapPinchDetector.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
-guid: aebd27fca271bde429fc692c3045f2d8
-timeCreated: 1455830843
+guid: 95ff52b4a0dd041479b65a44882d9e6f
+timeCreated: 1456448226
 licenseType: Pro
 MonoImporter:
   serializedVersion: 2
   defaultReferences: []
-  executionOrder: 1
+  executionOrder: -1001
   icon: {instanceID: 0}
   userData: 
   assetBundleName:
